Backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1

djangorestframework-simplejwt==5.3.1
openai>=1.0.0
django-ratelimit>=4.1.0
redis>=5.0.0
mysqlclient>=2.2.4
